# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Restaurante",
        "description": "API para la gestión de restaurantes",
        "contact": {
            "name": "Soporte API",
            "email": "soporte@restaurante.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/mesas": {
            "get": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Lista todas las mesas",
                "operationId": "c6cbe50a654a8b29e414c8631a2bafdf",
                "parameters": [
                    {
                        "name": "zona",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "capacidad",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de mesas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MesaCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Crea una nueva mesa",
                "operationId": "8388b6340022aa64251a0dd69734d8cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMesaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mesa creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MesaResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/mesas/{mesa}": {
            "get": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Muestra una mesa específica",
                "operationId": "fe3ca9a9095fbd0eb1302b1acf94ae54",
                "parameters": [
                    {
                        "name": "mesa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mesa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MesaResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mesa no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Actualiza una mesa existente",
                "operationId": "3328dad430eab185a8a379dbb9f1d682",
                "parameters": [
                    {
                        "name": "mesa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMesaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mesa actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MesaResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mesa no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Elimina una mesa",
                "operationId": "64c91f4f0a4bb46ef30972e815b946bd",
                "parameters": [
                    {
                        "name": "mesa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Mesa eliminada"
                    },
                    "404": {
                        "description": "Mesa no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/mesas/available": {
            "get": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Lista las mesas disponibles",
                "operationId": "d08e7792e0db8f5b31be72920f2be634",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "time",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "time"
                        }
                    },
                    {
                        "name": "guests",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de mesas disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MesaCollection"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Mesa": {
                "required": [
                    "numero",
                    "capacidad",
                    "estado"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "numero": {
                        "type": "integer",
                        "example": 1
                    },
                    "capacidad": {
                        "type": "integer",
                        "example": 4
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "disponible",
                            "ocupada",
                            "reservada"
                        ],
                        "example": "disponible"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "MesaResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Mesa"
                    }
                },
                "type": "object"
            },
            "StoreMesaRequest": {
                "required": [
                    "numero",
                    "capacidad"
                ],
                "properties": {
                    "numero": {
                        "type": "integer",
                        "example": 1
                    },
                    "capacidad": {
                        "type": "integer",
                        "example": 4
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "disponible",
                            "ocupada",
                            "reservada"
                        ],
                        "example": "disponible"
                    }
                },
                "type": "object"
            },
            "UpdateMesaRequest": {
                "properties": {
                    "numero": {
                        "type": "integer",
                        "example": 1
                    },
                    "capacidad": {
                        "type": "integer",
                        "example": 4
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "disponible",
                            "ocupada",
                            "reservada"
                        ],
                        "example": "disponible"
                    }
                },
                "type": "object"
            },
            "MesaCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Mesa"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string"
                            },
                            "last": {
                                "type": "string"
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer"
                            },
                            "from": {
                                "type": "integer"
                            },
                            "last_page": {
                                "type": "integer"
                            },
                            "path": {
                                "type": "string"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "to": {
                                "type": "integer"
                            },
                            "total": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}